requirements.txt
Django==2.0.5
django-registration==3.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
html5lib==1.0.1
httplib2==0.9.1
psycopg2==2.8.6
reportlab
django-crum
requests
gunicorn
